requirements.txt
discord.py
berserk
chess
python-dotenv